pyproject.toml
[project]
name = "langmem-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langmem>=0.0.29",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
]


